:truck: Move files to `bin/` and `lib/`
