+++ lib/common.ts
