+++ lib/euclidean.ts
