+++ lib/graph.ts
