+++ lib/standard.ts
