+++ lib/worker.ts
