+++ tsconfig.json
@@ -15,16 +15,16 @@
 		"target": "ES6",
 	},
   "files": [
-    "./aStar.ts",
-    "./common.ts",
-    "./coordinate.ts",
-    "./coordinateUnit.ts",
-    "./euclidean.ts",
     "./example.ts",
-    "./generate.ts",
-    "./graph.ts",
-    "./solve.ts",
-    "./standard.ts",
-    "./worker.ts",
+    "bin/generate.ts",
+    "bin/solve.ts",
+    "lib/aStar.ts",
+    "lib/common.ts",
+    "lib/coordinate.ts",
+    "lib/coordinateUnit.ts",
+    "lib/euclidean.ts",
+    "lib/graph.ts",
+    "lib/standard.ts",
+    "lib/worker.ts",
   ],
 }
